created photo-demo app
This is a "hello world" style app for the face API, to be released
to third parties via GitHub.  It runs face/landmark detection on a
photo and displays the photo with graphics indicating the landmark
positions.

Also updated the face tracker demo to remove the titlebar and change
to the new mobile vision icon.

Change-Id: I2a3c88a7c83efd1efa6b9b8d17b13e532daa7d70