open sourcing CameraSource
Please read over this carefully, including the comments.  This class will
be released to third party developers.

The reason for open sourcing this class is that third party developers
want more customization than this class is capable of exposing (e.g.,
support for torch mode, selecting a specific resolution, etc.).  Part
of the difficulty was that this class uses camera1 under the hood, but
we don't want to expose any details of camera1 in the API because it
is deprecated.  Another difficulty is that some developers want to use
features that aren't backward compatible, but we don't want to have
such features in the official API.  Open sourcing this class will give
developers the flexibility to customize this however it best suits
their particular app.

This class doesn't include any trade secrets, vision related or
otherwise.  But it is fairly detailed and hairy.

Change-Id: Ib2e0b0d2d3d0526b20783051718c9f48cfca50af